refactor: move read/write funcs to sdo service

+++ src/ucanopen/ucanopen_def.h
@@ -412,7 +412,7 @@ enum ODObjectAccessPermission {
 
 struct ODObjectKey {
     uint16_t index;
-    uint16_t subindex;
+    uint8_t subindex;
 };
